requirements.txt
# Core ML Framework (Blackwell RTX 50-series compatible)
--extra-index-url https://download.pytorch.org/whl/nightly/cu128
--pre
torch
torchvision

# Computer Vision (Phase 2)
ultralytics==8.3.50
opencv-python==4.10.0.84
pillow==11.0.0

# NLP & LLM (Phase 5)
transformers==4.47.1
peft==0.14.0
bitsandbytes==0.45.0       # ADDED: QLoRA quantization
datasets==3.2.0
accelerate==1.2.1
sentencepiece==0.2.0
protobuf==5.29.2           # ADDED: Required by transformers

# Web Interface (Phase 8)
streamlit==1.40.2

# Utilities
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.6.0
scipy==1.14.1              # ADDED: Required by peft/transformers
tqdm==4.67.1
pyarrow>=12.0.0
pyyaml>=6.0
requests>=2.31.0

# Training Monitoring (Phase 2)
tensorboard==2.18.0        # ADDED: Monitor training curves
matplotlib==3.9.3          # ADDED: Visualization (optional)

# Development
pytest==8.3.4
black==24.10.0
flake8==7.1.1